money number type set to Double everywhere Transfer min is now 1 Added placeholder texts for new fund form

+++ src/main/java/hu/progmasters/fundraiser/domain/Account.java
@@ -40,7 +40,7 @@ public class Account {
 
     @Min(0)
     @Column(name = "balance", nullable = false)
-    private Integer balance = 5000;
+    private Double balance = 5000.0;
 
     @Enumerated(EnumType.STRING)
     @ElementCollection(targetClass = AccountRole.class,
@@ -88,11 +88,11 @@ public class Account {
         this.password = password;
     }
 
-    public Integer getBalance() {
+    public Double getBalance() {
         return balance;
     }
 
-    public void setBalance(Integer balance) {
+    public void setBalance(Double balance) {
         this.balance = balance;
     }
